+++ src/blocks/homepage-articles/utils.ts
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { times, isEqual, isUndefined, pick, pickBy } from 'lodash';
+import { times, isEqual, isNull, isUndefined, pick, pickBy } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -40,6 +40,8 @@ const POST_QUERY_ATTRIBUTES = [
 	'postType',
 	'includedPostStatuses',
 	'deduplicate',
+	'showCaption',
+	'showCredit',
 ];
 
 /**
@@ -70,6 +72,8 @@ export const queryCriteriaFromAttributes = ( attributes: Block[ 'attributes' ] )
 		excerptLength,
 		postType,
 		showExcerpt,
+		showCaption,
+		showCredit,
 		tags,
 		customTaxonomies,
 		specificPosts = [],
@@ -91,12 +95,12 @@ export const queryCriteriaFromAttributes = ( attributes: Block[ 'attributes' ] )
 			  }
 			: {
 					postsToShow,
-					categories,
+					categories: validateAttributeCollection( categories ),
 					includeSubcategories,
-					authors,
-					tags,
-					tagExclusions,
-					categoryExclusions,
+					authors: validateAttributeCollection( authors ),
+					tags: validateAttributeCollection( tags ),
+					tagExclusions: validateAttributeCollection( tagExclusions ),
+					categoryExclusions: validateAttributeCollection( categoryExclusions ),
 					customTaxonomyExclusions,
 					customTaxonomies,
 					postType,
@@ -106,11 +110,16 @@ export const queryCriteriaFromAttributes = ( attributes: Block[ 'attributes' ] )
 	);
 	criteria.excerptLength = excerptLength;
 	criteria.showExcerpt = showExcerpt;
+	criteria.showCaption = showCaption;
+	criteria.showCredit = showCredit;
 	return criteria;
 };
 
 export const sanitizePostList = ( postList: HomepageArticlesAttributes[ 'specificPosts' ] ) =>
 	postList.map( id => parseInt( id ) ).filter( id => id > 0 );
+
+export const validateAttributeCollection = ( attr: Array< number > ) =>
+	pickBy( attr, ( value: unknown ) => ! isUndefined( value ) && ! isNull( value ) );
 
 /**
  * Each eligible block's attributes can be used to create a posts query.